+++ app/Http/Controllers/PengasuhController.php
@@ -4,12 +4,14 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Pengasuh;
+use App\Theme;
+use Illuminate\Support\Facades\Validator;
 
 class PengasuhController extends Controller
 {
     public function index()
     {
-        return view('admin.pengasuh');
+        return view('admin.pengasuh',['tema' => Theme::find(1), 'title' => 'Pengasuh']);
     }
 
 
@@ -23,6 +25,13 @@ class PengasuhController extends Controller
 
     public function update(Request $request)
     {
+        $validator = Validator::make($request->all(),[
+            'nama' => 'required',
+            'foto' => 'image|max:2048'
+        ]);
+        if ($validator->fails()) {
+            return $validator->errors()->all();
+        }
         $pengasuh = Pengasuh::find(1);
         if ($request->file('foto')) {
             \Storage::delete($pengasuh->foto);
@@ -30,15 +39,11 @@ class PengasuhController extends Controller
         } else {
             $foto = $pengasuh->foto;
         }
-        $request->validate([
-            'nama' => 'required',
-            'foto' => 'image|max:2048'
-        ]);
         Pengasuh::where('id', 1)->update([
             'nama' => $request->nama,
             'sambutan' => $request->sambutan,
             'foto' => $foto
         ]);
-        return $request;
+        return true;
     }
 }
