+++ app/Http/Controllers/TemaController.php
@@ -4,13 +4,14 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Theme;
+use Illuminate\Support\Facades\Validator;
 
 class TemaController extends Controller
 {
 
     public function index()
     {
-        return view('admin.tema');
+        return view('admin.tema',['tema' => Theme::find(1), 'title' => 'Tema']);
     }
 
 
@@ -24,6 +25,15 @@ class TemaController extends Controller
 
     public function update(Request $request)
     {
+        $validator = Validator::make($request->all(), [
+            'judul' => 'required',
+            'bg1' => ['image','max:2048'],
+            'logo' => ['image','max:2048'],
+        ]);
+
+        if ($validator->fails()) {
+            return $validator->errors()->all();
+        }
         $tema = Theme::find(1);
         if ($request->file('bg1')) {
             \Storage::delete($tema->bg1);
@@ -37,17 +47,22 @@ class TemaController extends Controller
         } else {
             $foto2 = $tema->bg2;
         }
-        $request->validate([
-            'judul' => 'required',
-            'tema' => 'required'
-        ]);
+        if ($request->file('logo')) {
+            \Storage::delete($tema->logo);
+            $logo = $request->file('logo')->store('tampilan');
+        } else {
+            $logo = $tema->logo;
+        }
+    
+
         Theme::where('id', 1)->update([
             'judul' => $request->judul,
+            'judul2' => $request->judul2,
             'bg1' => $foto1,
             'bg2' => $foto2,
-            'tema' => $request->tema
-        ]);
-        return $tema;
+            'logo' => $logo
+        ]); 
+        return true;
     }
 
 }
